monkey: eval: function and call
I'm not sure that `object` and `eval` packages should be splitted as the
reference book does. The `Environment` interfaces in object package
looks ad-hoc.

Q. Why did I need to define such an interface? The ref book does not
   have it.
A. That's why I had chose the `eval` package to contain the
   `Environment` _struct_).

Ref: chapter 3.10